add clear button instead of click to clear
