+++ src/main/java/mrriegel/storagenetwork/CraftingRecipes.java
@@ -1,15 +1,8 @@
 package mrriegel.storagenetwork;
-import java.util.List;
-import mrriegel.storagenetwork.remote.ItemRemote;
 import net.minecraft.init.Blocks;
 import net.minecraft.init.Items;
-import net.minecraft.inventory.InventoryCrafting;
 import net.minecraft.item.ItemStack;
-import net.minecraft.item.crafting.ShapelessRecipes;
 import net.minecraftforge.fml.common.registry.GameRegistry;
-import net.minecraftforge.oredict.RecipeSorter;
-import net.minecraftforge.oredict.RecipeSorter.Category;
-import net.minecraftforge.oredict.ShapedOreRecipe;
 
 public class CraftingRecipes {
   public static void init() {
