+++ src/main/java/mrriegel/storagenetwork/cable/BlockCable.java
@@ -6,12 +6,10 @@ import javax.annotation.Nullable;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import mrriegel.storagenetwork.AbstractBlockConnectable;
-import mrriegel.storagenetwork.ConfigHandler;
 import mrriegel.storagenetwork.CreativeTab;
 import mrriegel.storagenetwork.GuiHandler;
 import mrriegel.storagenetwork.IConnectable;
 import mrriegel.storagenetwork.ModBlocks;
-import mrriegel.storagenetwork.ModItems;
 import mrriegel.storagenetwork.StorageNetwork;
 import mrriegel.storagenetwork.cable.TileCable.Kind;
 import mrriegel.storagenetwork.helper.InvHelper;
@@ -25,7 +23,6 @@ import net.minecraft.client.resources.I18n;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.init.Blocks;
 import net.minecraft.item.ItemBlock;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
