+++ src/main/resources/assets/storagenetwork/lang/en_US.lang
@@ -29,6 +29,7 @@ gui.storagenetwork.fil.tooltip_IN=Only input
 gui.storagenetwork.fil.tooltip_OUT=Only output
 gui.storagenetwork.fil.tooltip_BOTH=Input and output
 
+gui.storagenetwork.fil.tooltip_clear=Clear Search
 
 tooltip.storagenetwork.indicator=Gives a redstone signal if more/less items are in the network.
